Fix tab view reload bug
Moved the SearchView tab view models to PokedexApp and passed them
through the environment. A boolean that tracks whether or not the
view has appeared was also added to the view models.

As a result, when a view has already appeared it no longer does a network call
when it first appears.