Refactor SearchByNameOrID functions to throw
SearchByNameOrID protocol functions will now throw instead of
returning nil. This was done so that the caller has more control,
and can see what went wrong.

Update all files that used the previous nil API to use try? at
call site. This makes the code same as it was before, but now has
the ability to show errors to user if needed.