Add new search functionality
The search bar has now been given a manager that stores the state of
the search bar. This enabled updates to the text to be debounced in
the onReceive function of the views that need it. SearchText was
removed because it was no longer useful and was redundant.

Extened Set to include containsItem(named:). This was done inorder
to remove reused code in the view models. Added the Namable protocol
in order to implement the containsItem(named:) function.