Update loading logic for SearchView tabs
Changed the initial loading to use the viewHasAppeared bool,
instead of isLoading. This was done to simplify the loading logic
when the view/tab first appears. It doesn't need to be reset
(like isLoading does) because the tabs do not reload when they
re-appear. Removed local loading view views to use a new view
named LoadingView.