Refactor MachinesListView's structure
Changed MachinesListView to look and function like other detail
helper views. Meaning that MachinesListView is now an extension
of MoveDetail.

Removed action property on PopoverNavigationBar and replaced it
with dismiss. This was done because dismiss always dismisses the
presented view and calling dismiss from the parent wasn't necessary.

Updated MoveDetail to only show a count of the machines instead of
displaying them. The details of the machines can now be found by
tapping on the show more button next to the count.

LoadingView has been moved to its own file.