Add pagination to AbilityPokemonListViewViewModel
Pagination was added to the AbilityDetail's PokemonListView.
This was done to help cut down on the initial load time. Even if
the data is cached, it will now load faster due to having less
stuff to load.

Removed the PokemonListView extension on AbilityDetail and renamed
the file to AbilityPokemonListView. Having less indentation is easier
to read and looks better.

Added SearchableByURL to help look up some PokeAPI data easier.