Refactor searching functionality
Refactored how searching works in the app. Using a little bit of
Combine allows the search text to be limited and filtered before
being used. This also removes the duplicated code from the view
models.

View models for the search view tabs have been moved to the HomeView.
This was done to make the view models only appear once and be kept
alive, without needing to be passed through the environment. (if
something is only going one level down, I prefer to use
ObserableObjects.)

Added some state to the view models to reduce the amount of if
statements in the views. Having state makes managing the state
of the view model easier (since it can only ever be in one state).

Refactored LoadingView UI to include a spinning image.
Added SearchErrorView.