Refactor PokemonTypeTag
PokemonTypeTag now takes the appropriate type not just a string name.
This makes the call sight much cleaner and clearer.

PokemonTypeTag now has a navigation link. PokemonTypeTag can now be
tapped from anywhere it appears. This reduces the amounts of
navigation links that need to written in the different places
PokemonTypeTag is seen.