Refactor view loading logic
Refactored the way views are loaded and view state is tracked.
Added a new ViewState enum to keep track of the states a view can be in.
Removed error prone "if viewAppeared ..." code and moved it into the
view model. This means the view model keeps track of and updates its own
state internally.

Replaced all instances of viewHasAppeared with a viewState.
Refactored if statements in views to switch statements.